rename host to cluster
also removes unrelevant data
