+++ metricbeat/docs/modules/couchbase.asciidoc
@@ -31,11 +31,11 @@ metricbeat.modules:
 
 The following metricsets are available:
 
-* <<metricbeat-metricset-couchbase-host,host>>
+* <<metricbeat-metricset-couchbase-cluster,cluster>>
 
 * <<metricbeat-metricset-couchbase-node,node>>
 
-include::couchbase/host.asciidoc[]
+include::couchbase/cluster.asciidoc[]
 
 include::couchbase/node.asciidoc[]
