--- metricbeat/docs/modules/couchbase/host.asciidoc
@@ -1,19 +1,0 @@
-////
-This file is generated! See scripts/docs_collector.py
-////
-
-[[metricbeat-metricset-couchbase-host]]
-include::../../../module/couchbase/host/_meta/docs.asciidoc[]
-
-
-==== Fields
-
-For a description of each field in the metricset, see the
-<<exported-fields-couchbase,exported fields>> section.
-
-Here is an example document generated by this metricset:
-
-[source,json]
-----
-include::../../../module/couchbase/host/_meta/data.json[]
-----
