+++ metricbeat/module/couchbase/cluster/cluster.go
@@ -1,4 +1,4 @@
-package host
+package cluster
 
 import (
 	"github.com/elastic/beats/libbeat/common"
@@ -23,7 +23,7 @@ var (
 // init registers the MetricSet with the central registry.
 // The New method will be called after the setup of the module and before starting to fetch data
 func init() {
-	if err := mb.Registry.AddMetricSet("couchbase", "host", New, hostParser); err != nil {
+	if err := mb.Registry.AddMetricSet("couchbase", "cluster", New, hostParser); err != nil {
 		panic(err)
 	}
 }
@@ -75,6 +75,5 @@ func (m *MetricSet) Fetch() (common.MapStr, error) {
 		return nil, fmt.Errorf("Error Connecting to Couchbase %d: %s", resp.StatusCode, resp.Status)
 	}
 
-
-	return eventMapping(resp.Body, m.Host()), nil
+	return eventMapping(resp.Body), nil
 }
