--- metricbeat/module/couchbase/host/_meta/docs.asciidoc
@@ -1,3 +1,0 @@
-=== couchbase host MetricSet
-
-This is the host metricset of the module couchbase.
